+++ packages/kbn-search-api-panels/tsconfig.json
@@ -23,6 +23,7 @@
     "@kbn/security-plugin",
     "@kbn/console-plugin",
     "@kbn/ui-theme",
-    "@kbn/try-in-console"
+    "@kbn/try-in-console",
+    "@kbn/test-jest-helpers"
   ]
 }
